feat(test-infra): Phase A — the three-arm database fixture, calibrated green
Selector truth table (TEST_DATABASE_URL BYO > TEST_DB=docker|neon > offline
default) with hard errors. OFFLINE arm: derived postgis+pgvector image
(digest-pinned amd64-only base + pgvector 0.8.5-1.pgdg11+1; emulated locally,
documented), clean template1 database (the postgis image pre-seeds tiger
schemas that atlas's clean-check refuses), sslmode=disable, pinned-atlas apply
— the pgvector neutralization and swallow-failures loop are gone. NEON arm:
neon_local manages BRANCH LIFECYCLE ONLY; asyncpg connects to the ephemeral
branch's DIRECT cloud endpoint (per Phase-0's proxy-wire rejection); teardown
does graceful SIGTERM then an API delete fallback (testcontainers force-kill
orphans branches — observed live, 4 strays swept). BYO arm: read-verify +
fail-on-behind + mutation opt-in + protected-lineage rejection. Shared atlas
helper (0.30.0 pin, --revisions-schema public, sanitized-stderr errors) also
fixes the provisioner's double-ledger bug. route_anime atlas twin + atlas.sum.
Eval runner honors TEST_DATABASE_URL under EVAL_FULLSTACK.

Maiden calibration caught and fixed REAL bugs never executed by CI (which runs
unit only): save_route's ST_MakePoint overload ambiguity (prepare-time failure
in production code), a uuid-vs-str test assertion, and a serialization-contract
split — the API now guarantees stable keys (explicit nulls) per the journey/
selected-route consumer contracts, with the unit expectation aligned.

Calibration: offline 65 passed in 29s; neon 66 passed in 6m21s (16x — the
local-default=offline verdict re-confirmed); unit 1107 green.

Restored from the codex session transcript after a tmp wipe (gate reproduced
the original 1107 exactly).

+++ apps/agent/agent/infrastructure/supabase/repositories/routes.py
@@ -46,8 +46,8 @@ class RoutesRepository:
                             total_distance, total_duration, route_data
                         )
                         VALUES ($1, $2,
-                                CASE WHEN $3 IS NOT NULL
-                                     THEN ST_MakePoint($3, $4)::geography
+                                CASE WHEN $3::float8 IS NOT NULL
+                                     THEN ST_MakePoint($3::float8, $4::float8)::geography
                                      ELSE NULL END,
                                 $5, $6, $7, $8::jsonb)
                         RETURNING id
